requirements.txt
PyYAML==6.0.1
elastic-transport==8.13.1
elasticsearch==8.14.0
requests==2.32.3
urllib3==2.2.2
feedparser==6.0.11
beautifulsoup4
openai
pytz
jinja2

